requirements.txt
streamlit>=1.45.0
google-genai>=1.0.0
python-dotenv>=1.0.0
Pillow>=11.0.0
PyJWT>=2.0.0
requests>=2.31.0
